requirements.base.txt
django>=4.1.7,<4.2.0
djangorestframework>=3.14.0,<3.15.0
psycopg2>=2.9.5,<3.0.0
drf-spectacular>=0.26.0,<0.27
Pillow>=9.4.0,<9.5.0